# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mediasearch",
    "summary": "EmbeddingGemma 2 を使った録画映像・音声・画像のマルチモーダル検索 API(開発用)",
    "description": "\n録画映像・音声・画像を取り込み、**文章・画像・音声**で該当する場面を探すための API です。\n\n- 取り込み(`/api/ingest/*`)はバックグラウンドのジョブで行い、応答の `job_id` で進み具合を確認します\n  (`GET /api/jobs/{job_id}`)。\n- 検索(`/api/search/*`)は、スコア(コサイン類似度)の高い順に窓(時間の区切り、または静止画 1 枚)を返します。\n- 元ファイルは `GET /api/media/{source_id}`(Range 対応)、サムネイルは `GET /api/thumb/{window_id}` で取得します。\n- エラーは `{\"detail\": \"...\"}` の形で返します(説明は日本語)。\n\n**認証**: サーバに `API_TOKEN` が設定されている場合、`Authorization: Bearer <トークン>` か\n`X-API-Key: <トークン>` が必要です\n(右上の Authorize から設定できます)。設定されていない場合は不要です。\n\nAI エージェントからは、同じ機能を MCP(`/mcp`)で利用できます。\n開発者向けの説明は docs/api.md と docs/mcp.md を参照してください。\n",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/info": {
      "get": {
        "tags": [
          "状態"
        ],
        "summary": "状態・版・既定値・索引の件数",
        "description": "推論側(モデル・デバイス)、アプリの版、取り込みの既定値とプリセット、索引の件数、監視フォルダの状態を返します。",
        "operationId": "info_api_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/ingest/video": {
      "post": {
        "tags": [
          "取り込み"
        ],
        "summary": "動画をアップロードして取り込む",
        "description": "動画を受け付け、バックグラウンドの取り込みジョブを作ります。\n\n進み具合は `GET /api/jobs/{job_id}` で確認します。",
        "operationId": "ingest_video_api_ingest_video_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_video_api_ingest_video_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DB のモデル・次元と、現在の設定が一致しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる(MAX_UPLOAD_MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "映像・音声・画像として扱えないファイル、または入力の検証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/ingest/audio": {
      "post": {
        "tags": [
          "取り込み"
        ],
        "summary": "音声をアップロードして取り込む",
        "description": "音声を区切りごとにベクトルにします(窓の種類は audio)。",
        "operationId": "ingest_audio_api_ingest_audio_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_audio_api_ingest_audio_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DB のモデル・次元と、現在の設定が一致しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる(MAX_UPLOAD_MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "映像・音声・画像として扱えないファイル、または入力の検証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/ingest/image": {
      "post": {
        "tags": [
          "取り込み"
        ],
        "summary": "画像をアップロードして取り込む(複数可)",
        "description": "静止画を取り込みます。画像 1 枚が 1 件の取り込み元になります(動画から切り出せない場合など)。\n\n1 枚ずつ受け付け、取り込めない画像があっても残りは処理します(結果は items と errors に分けて返します)。",
        "operationId": "ingest_image_api_ingest_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_image_api_ingest_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageBatchResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DB のモデル・次元と、現在の設定が一致しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる(MAX_UPLOAD_MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/ingest/path": {
      "post": {
        "tags": [
          "取り込み"
        ],
        "summary": "コンテナ内のファイルを取り込む",
        "description": "録画フォルダなどをマウントし、コンテナ内のパスで指定して取り込みます(アップロードより速い)。\n\n取り込み済みのファイルは重複として既存の取り込み元を返します(`force=true` で取り込み直し)。",
        "operationId": "ingest_path_api_ingest_path_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "許可していないフォルダ・別サイトからのリクエスト・ホスト名",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DB のモデル・次元と、現在の設定が一致しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "映像・音声・画像として扱えないファイル、または入力の検証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "description": "エラーの内容(日本語の説明。入力の検証エラーでは項目ごとの一覧)"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "examples": [
          {
            "detail": "取り込み元が見つかりません"
          }
        ]
      },
...
      "InfoResponse": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "version": {
            "$ref": "#/components/schemas/VersionInfo"
          },
          "embedder": {
            "$ref": "#/components/schemas/EmbedderStatus"
          },
          "defaults": {
            "additionalProperties": true,
            "type": "object",
            "title": "Defaults",
            "description": "取り込みの既定値(窓の長さなど)"
          },
          "presets": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Presets",
            "description": "取り込みのプリセット"
          },
          "index": {
            "$ref": "#/components/schemas/IndexStatus"
          },
          "watch": {
            "additionalProperties": true,
            "type": "object",
            "title": "Watch",
            "description": "監視フォルダの状態"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "role",
          "version",
          "embedder",
          "defaults",
          "presets",
          "index",
          "watch"
        ],
        "title": "InfoResponse"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_TOKEN の値"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API_TOKEN の値"
      }
    }
  },
...
}